Revert "am65-cpsw-nuss: fix inverted rgmii_id bit"
In 1017a61, rgmii_id was incorrectly assessed as inverted. In device
tree, phy-mode bindings "rgmii-id" and "rgmii-txid" refer to delays
added internally by the PHY, not by the MAC or the layout. The bit in
question was previously used on some CPSW implementations to turn on (0)
or off (1) the *MAC* delay. Therefore, the previous implementation was
proper and the change should be reverted.

This reverts commit 1017a614c4c3dba54fb0bb454880a7b4a03b70a3.

Signed-off-by: Ken Sloat <ken.s@variscite.com>