variscite: Fix missing include in common video code
The common video code is missing the inclusion of <env.h>. This is
mandatory since the functions within rely on various U-Boot env_ calls.

On some platforms, this can result in a bug due to an implicit
declaration of env_get. When a compiler detects an implicit declaration,
it will generate the signature of the function. If this signature does
not match the true signature of the function, this can result in
unpredictable behavior. In systems affected by this bug, it manifested
itself by returning a pointer to the environment that was padded with
bogus 0xFs in the upper portion of the address. This resulted in access
to an address that was not correct and triggered a memory fault crashing
the system.

To fix this, simply include <env.h> as it ought to be.

Signed-off-by: Ken Sloat <ken.s@variscite.com>