RND-2157: k3-am625-var-som-symphony: align cpsw3g node with TI
In moving from U-Boot 2021.01 to 2023.04, TI refactored cpsw3g node, by
moving the mdio pinctrl to the cpsw3g_mdio node.

TI appears to have rebased on this upstream, so there is no single commit
to observe this isolated change in U-Boot, however the difference can be
seen by doing a git diff of the U-Boot versions themselves.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>