RND-2160: am62x_var_som.env: define main_cpsw0_qsgmii_phyinit
The am62x_var_som.env file includes ti_armv7_common.env, which sets up
the common environment for TI boards based on the ARMv7 architecture.
Within the TI environment file, the following command is executed:

  #if CONFIG_CMD_REMOTEPROC
  	run main_cpsw0_qsgmii_phyinit; run boot_rprocs;
  #endif

However, the main_cpsw0_qsgmii_phyinit variable is expected to be defined
elsewhere, and if it is not, the following error message is displayed
during boot:

  Error: "main_cpsw0_qsgmii_phyinit" not defined

To avoid this error message, define main_cpsw0_qsgmii_phyinit with a dummy
value.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>