AM62: Secure Boot: Adapt Device Trees for Secure Boot
This is a part of the port of the Release 9.0 for Secure Boot from
ti-u-boot-2023.04_09.02.01.10_var01

This is based on the following commits:

6ee82c2: am62x_var_som: provide DeviceTrees for Secure Boot and non-Secure Boot

This allows an easy shifting between building for HS-SE (security-enabled = Secure Boot) devices and HS-FS (field-securable, not yet fused),
especially when working with Yocto builds.

This is specifically relevant when building for the R5 core which is the target for the tiboot3.bin, being the first customer-build binary.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>