am62x_var_som_secureboot: provide own build-in environment for Secure Boot
This provides an environment file for Secure Boot.

In its intial version is identical to the non-secure boot version, with the exception of forcing a boot from a FIT_Image (boot_fit = 1).

The file is called secureboot_prep as it it supposed to remind customers that this is provided as a reference only.
Customers must evaluate and modify the implementation as needed to align with their own security policies and deployment needs.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>