AM62x: Secure Boot: secureboot_prep defconfigs
This provieds the defconfig files used for Secure Boot.

The defconfigs were created with the following process:

- take the defconfig from the non-secure variant of U-Boot 2025.01 as a basis
- do a menuconfig to remove unsecure features (see below)
- run savedefconfig to generate the file
- Copy the resulting file

This defconfig is identical to the default file, but it removes the
CONFIGS_DISTRO_DEFAULTS feature, thereby making other features
selectable which would otherwise be selected by default and therefore
not visible in the config at all.
In general the following changes are made:

- remove support for booting Linux kernel images directly (booti
command) as this would pypass the fitImage signature check
- remove support for booting ELF images
- remove support for go command which would allow to directly jump to an
arbitrary address
- remove support for booting an EFI image

The remaining features are left for easier compatibility for now.
Further adaptions can be made to make restrictions stronger.

This defconfig does NOT YET set CONFIG_ENV_IS_NOWHERE for compatibility
reasons with system update mechanisms.

This is provided as a reference example. Customers must evaluate and
modify the implementation as needed to align with their own security
policies and deployment needs.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>
(cherry picked from commit 4f0b47d9ec588639a023854ce94095458251d086)