arm: dts: am62p5-var-som: allow to read back GPIO value
The AM62PX pin mux requires PIN_INPUT as a way to be able to read back
the value of a GPIO pin.
PIN_OUTPUT is not needed to drive the pin, PIN_INPUT also activates the
output driver.

Use PIN_INPUT to make certain helper functionality display correct
results (regulator U-Boot command, GPIO U-Boot command).
This does not have an affect on the actual pin value, but can help in
verification.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>